perf: :zap: Refactoring path pacakage and new stucture project for better state management
