+++ lib/app/components/bird.dart
@@ -4,9 +4,9 @@ import 'package:flame/effects.dart';
 import 'package:flame_audio/flame_audio.dart';
 import 'package:flappy_bird/app/components/style/image_local.dart';
 import 'package:flappy_bird/app/components/style/sound_local.dart';
-import 'package:flappy_bird/app/modules/game_start/controllers/configuration.dart';
-import 'package:flappy_bird/app/modules/game_start/controllers/movement.dart';
-import 'package:flappy_bird/app/modules/game_start/views/GamePlay.dart';
+import 'package:flappy_bird/app/modules/game_start/movement.dart';
+import 'package:flappy_bird/app/modules/game_start/configuration.dart';
+import 'package:flappy_bird/app/modules/game_start/GamePlay.dart';
 import 'package:flutter/material.dart';
 
 class Bird extends SpriteGroupComponent<BirdMovement>
