+++ lib/app/components/ground.dart
@@ -5,8 +5,8 @@ import 'package:flame/components.dart';
 import 'package:flame/flame.dart';
 import 'package:flame/parallax.dart';
 import 'package:flappy_bird/app/components/style/image_local.dart';
-import 'package:flappy_bird/app/modules/game_start/controllers/configuration.dart';
-import 'package:flappy_bird/app/modules/game_start/views/GamePlay.dart';
+import 'package:flappy_bird/app/modules/game_start/configuration.dart';
+import 'package:flappy_bird/app/modules/game_start/GamePlay.dart';
 
 class Ground extends ParallaxComponent<GamePlay> with HasGameRef<GamePlay> {
   Ground();
