+++ lib/app/components/pipe.dart
@@ -4,9 +4,9 @@ import 'package:flame/collisions.dart';
 import 'package:flame/components.dart';
 import 'package:flame/flame.dart';
 import 'package:flappy_bird/app/components/style/image_local.dart';
-import 'package:flappy_bird/app/modules/game_start/controllers/configuration.dart';
-import 'package:flappy_bird/app/modules/game_start/controllers/pipe_position.dart';
-import 'package:flappy_bird/app/modules/game_start/views/GamePlay.dart';
+import 'package:flappy_bird/app/modules/game_start/configuration.dart';
+import 'package:flappy_bird/app/modules/game_start/pipe_position.dart';
+import 'package:flappy_bird/app/modules/game_start/GamePlay.dart';
 
 class Pipe extends SpriteComponent with HasGameRef<GamePlay> {
   Pipe({required this.pipePosition, required this.height});
