+++ lib/app/components/pipe_ground.dart
@@ -5,9 +5,9 @@ import 'package:flame/components.dart';
 import 'package:flame_audio/flame_audio.dart';
 import 'package:flappy_bird/app/components/pipe.dart';
 import 'package:flappy_bird/app/components/style/sound_local.dart';
-import 'package:flappy_bird/app/modules/game_start/controllers/configuration.dart';
-import 'package:flappy_bird/app/modules/game_start/controllers/pipe_position.dart';
-import 'package:flappy_bird/app/modules/game_start/views/GamePlay.dart';
+import 'package:flappy_bird/app/modules/game_start/configuration.dart';
+import 'package:flappy_bird/app/modules/game_start/pipe_position.dart';
+import 'package:flappy_bird/app/modules/game_start/GamePlay.dart';
 import 'package:flutter/material.dart';
 
 class PipeGround extends PositionComponent with HasGameRef<GamePlay> {
